test: add replication/applier-rollback
Test that diag_raise doesn't happen if async transaction
fails inside replication procedure.

Side note: I don't like merging tests with patches in
general and I hate doing so for big tests with a passion
because it hides the patch code itself. So here is a
separate patch on top of the fix.

Test-of #4730

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>